authorize the user

+++ controllers/expense.js
@@ -2,33 +2,47 @@ const expenses=require('../model/expense');
 
 const sequelize = require('../util/database');
 
+function isInValid(string){
+    if(string===undefined || string.length===0){
+        return true
+    }else{
+        return false
+    }
+}
+
 exports.postExpense=async(req,res,next)=>{
     try{
         const {expenseAmount,description,category}=req.body;
-        const data=await expenses.create({expenseAmount,description,category})
-        res.status(201).json({newExpenseDetails:data})
+        if(isInValid(expenseAmount) || isInValid(description) || isInValid(category)){
+            res.status(400).json({message:'Missing parameters !',success:false});
+        }
+        const data=await req.user.createExpense({expenseAmount,description,category})
+        res.status(201).json({newExpenseDetails:data,success:true})
     }catch(err){
-        res.status(500).json({error:err}) 
+        res.status(500).json({error:err,success:false}) 
     }
 }
 
 exports.getExpenses=async(req,res,next)=>{
     try{
-        const allexpenses= await expenses.findAll();
-        res.status(200).json({allExpenses: allexpenses});
+        const allexpenses= await req.user.getExpenses()
+        res.status(200).json({allExpenses: allexpenses,success:true});
     }catch(err){
-        res.sendStatus(500).json({error:err});
+        res.sendStatus(500).json({error:err,success:false});
     }
 }
 
 exports.deleteExpense=async(req,res,next)=>{
     try{
-        if(req.params.id==='undefined'){
-            return res.status(400).json({err:'id not found'})
+        if(req.params.id==='undefined' || req.params.id.length===0 || req.user.id==='undefined' || req.user.id.length===0){
+            return res.status(400).json({err:'id not found',success:false})
         }
         const expenseid=req.params.id
-        await expenses.destroy({where:{id:expenseid}});
-        res.sendStatus(200);
+        const deleteexpense=await expenses.destroy({where:{id:expenseid,userId:req.user.id}});
+        if(deleteexpense===0){
+            return res.status(404).json({success:false,message:'Expense not belong to user'})
+        }
+        res.status(200).json({message:'deleted successfully',success:true});
     }catch(err){
         res.status(500).json({error:err});
     }
